refactor(pptx): extract master.rs from lib.rs (verbatim)
Master/layout inheritance: LayoutPlaceholders (struct + impl), InheritedBlipFill, the 11 master extractors (parse_master_anchors, MASTER_TXSTYLE_PH_TYPES, parse_master_alignments, parse_master_ea_ln_brk, parse_master_font_sizes, parse_master_level_font_sizes/indents/bullets, parse_master_txstyle_bold_italic/color/spacing, parse_master_transforms), parse_layout_placeholders, ParsedLayout (+ Default impl), read_show_master_sp, parse_layout, EffectiveMaster, build_master_bundle — now pub(crate) (incl. the LayoutPlaceholders lookup_* methods, called from lib.rs/shape.rs). master.rs imports fill/theme/text/shape helpers + shared XML helpers + note_layout_master_parse + ooxml_common blip. shape.rs now imports LayoutPlaceholders from crate::master; dead ooxml_common::blip import removed from lib.rs; lib.rs uses master::*.

Rename MasterBundle -> ParsedMaster (type name only, fields unchanged) in the same commit that creates master.rs.
