+++ packages/pptx/parser/src/master.rs
@@ -1,0 +1,1657 @@
+//! Slide-master / layout inheritance: the per-master extractors (anchors,
+//! alignments, ea-line-break, font sizes, per-level sizes/indents/bullets,
+//! txStyle bold/italic/colour/spacing, transforms), the layout-placeholder
+//! resolver, and the cached `ParsedMaster` (was `MasterBundle`) / `ParsedLayout`
+//! bundles. Extracted verbatim from `lib.rs`; the only non-move change is the
+//! `MasterBundle` → `ParsedMaster` type rename (fields unchanged).
+
+use crate::fill::{
+    parse_background, parse_blip_alpha, parse_color_node, parse_fill, parse_stroke, parse_xfrm,
+};
+use crate::shape::extract_decorative_shapes;
+use crate::text::{
+    empty_level_bullets, extract_level_bullets, extract_level_font_sizes, extract_level_indents,
+    extract_lvl1_font_size, has_any_level_bullet, has_any_level_indent, has_any_level_size,
+    merge_level_bullets, merge_level_indents, merge_level_sizes, read_level_bullets,
+    read_level_font_sizes, read_level_indents, LevelBullets, LevelFontSizes, LevelIndents,
+};
+use crate::theme::{bake_clr_map, parse_theme_colors};
+use crate::types::*;
+use crate::{
+    attr, attr_f64, attr_i64, attr_r, build_smartart_drawings, child, find_rel_target_by_type,
+    note_layout_master_parse, parse_rels, read_zip_str, resolve_path, PptxZip,
+};
+use ooxml_common::blip::{mime_from_ext, parse_src_rect, SrcRect};
+use std::collections::HashMap;
+
+/// Keyed first by idx (integer), then by type string.
+// `Clone` lets `parse_layout` cache one resolved `LayoutPlaceholders` per layout
+// and hand each slide a copy to layer its per-slide master txStyles fallbacks
+// onto without mutating the cached instance (D4).
+#[derive(Default, Clone)]
+pub(crate) struct LayoutPlaceholders {
+    pub(crate) by_idx: HashMap<u32, Transform>,
+    pub(crate) by_type: HashMap<String, Transform>,
+    /// Fallback transforms from slide master (by ph_type), used when layout has no xfrm
+    pub(crate) master_by_type: HashMap<String, Transform>,
+    /// Default font size (pt) per placeholder idx, from layout/master lstStyle
+    pub(crate) by_idx_font_size: HashMap<u32, f64>,
+    /// Default font size (pt) per placeholder type, from layout/master lstStyle
+    pub(crate) by_type_font_size: HashMap<String, f64>,
+    /// Per-list-level default font sizes (pt) per placeholder idx — index 0..=8
+    /// maps to lvl1pPr..lvl9pPr (ECMA-376 §21.1.2.4). Lets nested bullets shrink
+    /// per level (e.g. body 28pt → lvl2 24pt → lvl3 20pt) instead of all using
+    /// the level-1 size. None per level where the style chain doesn't specify it.
+    pub(crate) by_idx_level_sizes: HashMap<u32, LevelFontSizes>,
+    /// Per-list-level default font sizes (pt) per placeholder type.
+    pub(crate) by_type_level_sizes: HashMap<String, LevelFontSizes>,
+    /// Per-list-level paragraph indents (`marL`/`marR`/`indent`, EMU) per
+    /// placeholder idx — what a paragraph with no own `marL`/`marR`/`indent`
+    /// inherits from the authored list-style cascade (ECMA-376 §21.1.2.4.13),
+    /// used as the fallback before PowerPoint's hardcoded implicit defaults.
+    pub(crate) by_idx_level_indents: HashMap<u32, LevelIndents>,
+    /// Per-list-level paragraph indents per placeholder type.
+    pub(crate) by_type_level_indents: HashMap<String, LevelIndents>,
+    /// Per-list-level inherited bullet (buChar/buAutoNum/buNone) per placeholder
+    /// idx — what a paragraph with no explicit bullet inherits (ECMA-376 §19.7.10).
+    pub(crate) by_idx_level_bullets: HashMap<u32, LevelBullets>,
+    /// Per-list-level inherited bullet per placeholder type.
+    pub(crate) by_type_level_bullets: HashMap<String, LevelBullets>,
+    /// Default bold per placeholder type, from layout lstStyle defRPr b attribute
+    pub(crate) by_type_bold: HashMap<String, bool>,
+    /// Default italic per placeholder type, from layout lstStyle defRPr i attribute
+    pub(crate) by_type_italic: HashMap<String, bool>,
+    /// Default caps ("all"/"small") per placeholder type, from layout/master
+    /// lstStyle defRPr cap attribute (ECMA-376 §21.1.2.3.13)
+    pub(crate) by_type_caps: HashMap<String, String>,
+    /// Vertical anchor ("t"/"ctr"/"b") per placeholder type, from layout/master bodyPr
+    pub(crate) by_type_anchor: HashMap<String, String>,
+    /// Default paragraph alignment per placeholder type, from layout/master lstStyle
+    pub(crate) by_type_alignment: HashMap<String, String>,
+    /// Paragraph alignment per placeholder idx — layout placeholder's own algn,
+    /// falling back to the master per-type alignment. Checked before the
+    /// type-keyed maps so a body placeholder resolves to its OWN idx's style,
+    /// not an unrelated typeless placeholder (ECMA-376 §19.3.1.x idx matching).
+    pub(crate) by_idx_alignment: HashMap<u32, String>,
+    /// Default East Asian line-break (eaLnBrk) per placeholder type, from the
+    /// layout lstStyle > lvl1pPr @eaLnBrk (ECMA-376 §21.1.2.2.7)
+    pub(crate) by_type_ea_ln_brk: HashMap<String, bool>,
+    /// Default space-before (hundredths of pt) per placeholder type, from layout lstStyle
+    pub(crate) by_type_space_before: HashMap<String, i64>,
+    /// Default space-after (hundredths of pt) per placeholder type, from layout lstStyle
+    pub(crate) by_type_space_after: HashMap<String, i64>,
+    /// Default space-before from master txStyles (fallback when layout has none)
+    pub(crate) by_type_master_space_before: HashMap<String, i64>,
+    /// Default space-after from master txStyles (fallback when layout has none)
+    pub(crate) by_type_master_space_after: HashMap<String, i64>,
+    /// Stroke per placeholder type from layout spPr > ln
+    pub(crate) by_type_stroke: HashMap<String, Stroke>,
+    /// Stroke per placeholder idx from layout spPr > ln
+    pub(crate) by_idx_stroke: HashMap<u32, Stroke>,
+    /// Default line spacing (spcPct val, e.g. 90000 = 90%) per placeholder idx, from layout lstStyle
+    pub(crate) by_idx_line_spacing: HashMap<u32, f64>,
+    /// Default line spacing (spcPct val) per placeholder type, from layout lstStyle
+    pub(crate) by_type_line_spacing: HashMap<String, f64>,
+    /// Paragraph alignment per placeholder type from master lstStyle > lvl1pPr algn (fallback)
+    pub(crate) by_type_master_alignment: HashMap<String, String>,
+    /// East Asian line-break per placeholder type from master lstStyle > lvl1pPr
+    /// @eaLnBrk (fallback when the layout has none) — ECMA-376 §21.1.2.2.7
+    pub(crate) by_type_master_ea_ln_brk: HashMap<String, bool>,
+    /// Default line spacing from master txStyles (fallback when layout has none)
+    pub(crate) by_type_master_line_spacing: HashMap<String, f64>,
+    /// Inherited blipFill (data URL + src rect) per placeholder idx from layout spPr
+    pub(crate) by_idx_blip_fill: HashMap<u32, InheritedBlipFill>,
+    /// Inherited blipFill per placeholder type from layout spPr
+    pub(crate) by_type_blip_fill: HashMap<String, InheritedBlipFill>,
+    /// Default text color per placeholder idx, from layout lstStyle defRPr solidFill
+    pub(crate) by_idx_color: HashMap<u32, String>,
+    /// Default text color per placeholder type, from layout lstStyle defRPr solidFill
+    pub(crate) by_type_color: HashMap<String, String>,
+    /// Default text color from master (txStyles + spTree lstStyle) — fallback when layout has none
+    pub(crate) by_type_master_color: HashMap<String, String>,
+    /// `<p:spPr><a:solidFill | a:noFill | a:gradFill | a:pattFill>` per placeholder idx.
+    /// Used to inherit a layout-level shape fill (e.g. a tinted body placeholder)
+    /// onto slide-level shapes whose `<p:spPr>` is empty.
+    pub(crate) by_idx_fill: HashMap<u32, Fill>,
+    /// Same as `by_idx_fill` but keyed by placeholder type (fallback when idx
+    /// doesn't match a layout shape).
+    pub(crate) by_type_fill: HashMap<String, Fill>,
+}
+
+#[derive(Debug, Clone)]
+pub(crate) struct InheritedBlipFill {
+    /// Embedded zip path of the inherited picture-placeholder blip.
+    pub(crate) image_path: String,
+    /// MIME of the blip at `image_path`.
+    pub(crate) mime_type: String,
+    pub(crate) src_rect: Option<SrcRect>,
+    pub(crate) alpha: Option<f64>,
+}
+
+impl LayoutPlaceholders {
+    pub(crate) fn lookup(&self, ph_type: &str, ph_idx: Option<u32>) -> Option<&Transform> {
+        ph_idx
+            .and_then(|i| self.by_idx.get(&i))
+            .or_else(|| self.by_type.get(ph_type))
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type.get("")
+                } else {
+                    None
+                }
+            })
+            .or_else(|| self.master_by_type.get(ph_type))
+    }
+
+    /// Look up the inherited default font size for a placeholder (layout then master fallback).
+    /// Idx-strict per ECMA-376 §19.7.16 (see `lookup_fill`'s rationale).
+    pub(crate) fn lookup_font_size(&self, ph_type: &str, ph_idx: Option<u32>) -> Option<f64> {
+        if let Some(i) = ph_idx {
+            return self.by_idx_font_size.get(&i).copied();
+        }
+        self.by_type_font_size.get(ph_type).copied().or_else(|| {
+            if ph_type == "body" {
+                self.by_type_font_size.get("").copied()
+            } else {
+                None
+            }
+        })
+    }
+
+    /// Per-list-level inherited default font sizes (lvl1..lvl9). Same idx-strict
+    /// resolution as `lookup_font_size`. All-None when the placeholder has no
+    /// per-level styling.
+    pub(crate) fn lookup_level_font_sizes(
+        &self,
+        ph_type: &str,
+        ph_idx: Option<u32>,
+    ) -> LevelFontSizes {
+        if let Some(i) = ph_idx {
+            return self
+                .by_idx_level_sizes
+                .get(&i)
+                .copied()
+                .unwrap_or([None; 9]);
+        }
+        self.by_type_level_sizes
+            .get(ph_type)
+            .copied()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_level_sizes.get("").copied()
+                } else {
+                    None
+                }
+            })
+            .unwrap_or([None; 9])
+    }
+
+    /// Per-list-level inherited paragraph indents (lvl1..lvl9). Same idx-strict
+    /// resolution as `lookup_level_font_sizes`. All-default (every axis None) when
+    /// the placeholder has no authored per-level indent.
+    pub(crate) fn lookup_level_indents(&self, ph_type: &str, ph_idx: Option<u32>) -> LevelIndents {
+        if let Some(i) = ph_idx {
+            return self
+                .by_idx_level_indents
+                .get(&i)
+                .copied()
+                .unwrap_or_default();
+        }
+        self.by_type_level_indents
+            .get(ph_type)
+            .copied()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_level_indents.get("").copied()
+                } else {
+                    None
+                }
+            })
+            .unwrap_or_default()
+    }
+
+    /// Per-list-level inherited bullets (lvl1..lvl9). Same idx-strict resolution as
+    /// `lookup_level_font_sizes`. All-None when the placeholder inherits no bullet.
+    pub(crate) fn lookup_level_bullets(&self, ph_type: &str, ph_idx: Option<u32>) -> LevelBullets {
+        if let Some(i) = ph_idx {
+            return self
+                .by_idx_level_bullets
+                .get(&i)
+                .cloned()
+                .unwrap_or_else(empty_level_bullets);
+        }
+        self.by_type_level_bullets
+            .get(ph_type)
+            .cloned()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_level_bullets.get("").cloned()
+                } else {
+                    None
+                }
+            })
+            .unwrap_or_else(empty_level_bullets)
+    }
+
+    /// Look up inherited bold for this placeholder type.
+    pub(crate) fn lookup_bold(&self, ph_type: &str) -> Option<bool> {
+        self.by_type_bold.get(ph_type).copied().or_else(|| {
+            if ph_type == "body" {
+                self.by_type_bold.get("").copied()
+            } else {
+                None
+            }
+        })
+    }
+
+    /// Look up inherited italic for this placeholder type.
+    pub(crate) fn lookup_italic(&self, ph_type: &str) -> Option<bool> {
+        self.by_type_italic.get(ph_type).copied().or_else(|| {
+            if ph_type == "body" {
+                self.by_type_italic.get("").copied()
+            } else {
+                None
+            }
+        })
+    }
+
+    /// Look up inherited caps ("all"/"small") for this placeholder type.
+    pub(crate) fn lookup_caps(&self, ph_type: &str) -> Option<String> {
+        self.by_type_caps.get(ph_type).cloned().or_else(|| {
+            if ph_type == "body" {
+                self.by_type_caps.get("").cloned()
+            } else {
+                None
+            }
+        })
+    }
+
+    /// Look up inherited vertical anchor for this placeholder type.
+    pub(crate) fn lookup_anchor(&self, ph_type: &str) -> Option<String> {
+        self.by_type_anchor.get(ph_type).cloned().or_else(|| {
+            if ph_type == "body" {
+                self.by_type_anchor.get("").cloned()
+            } else {
+                None
+            }
+        })
+    }
+
+    /// Look up inherited paragraph alignment for this placeholder.
+    ///
+    /// A placeholder identified by `idx` resolves through its own slot
+    /// (`by_idx_alignment`), which `parse_layout_placeholders` pre-seeds with the
+    /// master per-type default. Unlike `lookup_fill`, falling through to the
+    /// type map on an idx miss is intentional and safe (the seed already encodes
+    /// the master tier) — but the `""` (typeless) fallback is gated to
+    /// `ph_idx.is_none()` so an idx/typed placeholder never borrows an unrelated
+    /// typeless sibling's alignment (ECMA-376 §19.3.1.36 idx matching).
+    pub(crate) fn lookup_alignment(&self, ph_type: &str, ph_idx: Option<u32>) -> Option<String> {
+        if let Some(i) = ph_idx {
+            if let Some(a) = self.by_idx_alignment.get(&i) {
+                return Some(a.clone());
+            }
+        }
+        // The `""` fallback represents a typeless (idx-less, body-category)
+        // placeholder; only a placeholder that is itself typeless may use it.
+        let allow_empty = ph_idx.is_none() && ph_type == "body";
+        self.by_type_alignment
+            .get(ph_type)
+            .cloned()
+            .or_else(|| {
+                if allow_empty {
+                    self.by_type_alignment.get("").cloned()
+                } else {
+                    None
+                }
+            })
+            .or_else(|| self.by_type_master_alignment.get(ph_type).cloned())
+            .or_else(|| {
+                if allow_empty {
+                    self.by_type_master_alignment.get("").cloned()
+                } else {
+                    None
+                }
+            })
+    }
+
+    // ECMA-376 §21.1.2.2.7 eaLnBrk inheritance, mirroring lookup_alignment:
+    // layout per-type → layout generic ("") for body → master per-type →
+    // master generic. None means no ancestor specified it (parse_paragraph then
+    // applies the spec default of true).
+    pub(crate) fn lookup_ea_ln_brk(&self, ph_type: &str) -> Option<bool> {
+        self.by_type_ea_ln_brk
+            .get(ph_type)
+            .copied()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_ea_ln_brk.get("").copied()
+                } else {
+                    None
+                }
+            })
+            .or_else(|| self.by_type_master_ea_ln_brk.get(ph_type).copied())
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_master_ea_ln_brk.get("").copied()
+                } else {
+                    None
+                }
+            })
+    }
+
+    pub(crate) fn lookup_space_before(&self, ph_type: &str) -> Option<i64> {
+        self.by_type_space_before
+            .get(ph_type)
+            .copied()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_space_before.get("").copied()
+                } else {
+                    None
+                }
+            })
+            .or_else(|| self.by_type_master_space_before.get(ph_type).copied())
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_master_space_before.get("").copied()
+                } else {
+                    None
+                }
+            })
+    }
+
+    pub(crate) fn lookup_space_after(&self, ph_type: &str) -> Option<i64> {
+        self.by_type_space_after
+            .get(ph_type)
+            .copied()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_space_after.get("").copied()
+                } else {
+                    None
+                }
+            })
+            .or_else(|| self.by_type_master_space_after.get(ph_type).copied())
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_master_space_after.get("").copied()
+                } else {
+                    None
+                }
+            })
+    }
+
+    /// Look up inherited blipFill from the layout placeholder spPr. Used when a slide
+    /// references a picture placeholder (e.g. ph type="pic") without its own blipFill —
+    /// the image defined on the layout's matching placeholder should render through.
+    /// Idx-strict per ECMA-376 §19.7.16 (see `lookup_fill`'s rationale).
+    pub(crate) fn lookup_blip_fill(
+        &self,
+        ph_type: &str,
+        ph_idx: Option<u32>,
+    ) -> Option<InheritedBlipFill> {
+        if let Some(i) = ph_idx {
+            return self.by_idx_blip_fill.get(&i).cloned();
+        }
+        self.by_type_blip_fill.get(ph_type).cloned()
+    }
+
+    /// Look up inherited stroke from the layout placeholder spPr > ln.
+    /// Idx-strict per ECMA-376 §19.7.16 (see `lookup_fill`'s rationale).
+    pub(crate) fn lookup_stroke(&self, ph_type: &str, ph_idx: Option<u32>) -> Option<Stroke> {
+        if let Some(i) = ph_idx {
+            return self.by_idx_stroke.get(&i).cloned();
+        }
+        self.by_type_stroke.get(ph_type).cloned().or_else(|| {
+            if ph_type == "body" {
+                self.by_type_stroke.get("").cloned()
+            } else {
+                None
+            }
+        })
+    }
+
+    /// Look up inherited default text color for this placeholder (layout then master fallback).
+    ///
+    /// The *layout* tier is idx-strict per ECMA-376 §19.7.16: when the slide-level
+    /// placeholder carries an explicit `idx`, a layout colour is inherited only from the
+    /// layout shape with the SAME idx — never a sibling body placeholder at a different
+    /// idx (which would leak an unrelated region's colour).
+    ///
+    /// The *master* `txStyles` tier (titleStyle/bodyStyle/otherStyle), however, is a
+    /// document-wide default keyed by placeholder *type* (§21.1.2.4 / §19.3.1) and is
+    /// inherited regardless of idx. So when the idx-matched layout shape defines no
+    /// colour, resolution must still fall through to `by_type_master_color`. Without
+    /// this, a body placeholder whose layout shape sets size-but-not-colour resolves to
+    /// no colour at all and the renderer defaults to black — instead of the master
+    /// bodyStyle colour (e.g. `schemeClr bg1` = white on a dark theme). (sample-9 slide 2+)
+    pub(crate) fn lookup_color(&self, ph_type: &str, ph_idx: Option<u32>) -> Option<String> {
+        if let Some(i) = ph_idx {
+            if let Some(c) = self.by_idx_color.get(&i) {
+                return Some(c.clone());
+            }
+            // Layout idx had no colour → fall through to the master type-keyed default.
+            return self.by_type_master_color.get(ph_type).cloned().or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_master_color.get("").cloned()
+                } else {
+                    None
+                }
+            });
+        }
+        self.by_type_color
+            .get(ph_type)
+            .cloned()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_color.get("").cloned()
+                } else {
+                    None
+                }
+            })
+            .or_else(|| self.by_type_master_color.get(ph_type).cloned())
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_master_color.get("").cloned()
+                } else {
+                    None
+                }
+            })
+    }
+
+    /// Look up the inherited shape fill from the layout placeholder's `<p:spPr>`.
+    /// Used when the slide-level shape leaves `<p:spPr>` empty (or with no fill
+    /// elements) and is bound to a placeholder.
+    ///
+    /// ECMA-376 §19.7.16 (placeholder inheritance) is asymmetric: when the
+    /// slide-level shape declares `<p:ph idx="N">` it is bound to *that*
+    /// specific layout slot — the only valid inheritance source is the layout
+    /// shape with idx=N. Falling back to `by_type_fill` here would let a
+    /// sibling body placeholder (a different idx, different region of the
+    /// layout) bleed its fill onto a placeholder that the spec says should
+    /// have no fill. This is exactly what regressed sample-2 slide-4: layout10
+    /// has `body[idx=12]` (header, no fill) and `body[idx=13]` (bullet box,
+    /// gray fill) — the type fallback was leaking the bullet box's gray onto
+    /// the header.
+    ///
+    /// The type-only fallback only applies when the slide-level shape itself
+    /// has no idx, in which case "first body placeholder we found" is the
+    /// best we can do.
+    pub(crate) fn lookup_fill(&self, ph_type: &str, ph_idx: Option<u32>) -> Option<Fill> {
+        if let Some(i) = ph_idx {
+            return self.by_idx_fill.get(&i).cloned();
+        }
+        self.by_type_fill.get(ph_type).cloned().or_else(|| {
+            if ph_type == "body" {
+                self.by_type_fill.get("").cloned()
+            } else {
+                None
+            }
+        })
+    }
+
+    /// Look up inherited line spacing (spcPct val, e.g. 90000 = 90%) for this placeholder.
+    /// Idx-strict per ECMA-376 §19.7.16 (see `lookup_fill`'s rationale).
+    pub(crate) fn lookup_line_spacing(&self, ph_type: &str, ph_idx: Option<u32>) -> Option<f64> {
+        if let Some(i) = ph_idx {
+            return self.by_idx_line_spacing.get(&i).copied();
+        }
+        self.by_type_line_spacing
+            .get(ph_type)
+            .copied()
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_line_spacing.get("").copied()
+                } else {
+                    None
+                }
+            })
+            .or_else(|| self.by_type_master_line_spacing.get(ph_type).copied())
+            .or_else(|| {
+                if ph_type == "body" {
+                    self.by_type_master_line_spacing.get("").copied()
+                } else {
+                    None
+                }
+            })
+    }
+}
+
+/// Parse bodyPr anchor ("t"/"ctr"/"b") from master placeholder shapes.
+///
+/// Takes the already-parsed master root element (`<p:sldMaster>`) so
+/// `build_master_bundle` can parse the master XML once and share the
+/// `Document` across every `parse_master_*` extractor (ECMA-376 §19.3.1.42).
+pub(crate) fn parse_master_anchors(root: roxmltree::Node<'_, '_>) -> HashMap<String, String> {
+    let mut map = HashMap::new();
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            let ph_node = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph");
+            if let Some(ph) = ph_node {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(anchor) = child(sp, "txBody")
+                    .and_then(|tb| child(tb, "bodyPr"))
+                    .and_then(|bp| attr(&bp, "anchor"))
+                {
+                    map.entry(ph_type).or_insert(anchor.to_string());
+                }
+            }
+        }
+    }
+    map
+}
+
+/// txStyles style node → the placeholder types it defaults. ECMA-376 §19.3.1.52
+/// txStyles → titleStyle §19.3.1.49 / bodyStyle §19.3.1.5 / otherStyle §19.3.1.35.
+pub(crate) const MASTER_TXSTYLE_PH_TYPES: &[(&str, &[&str])] = &[
+    ("titleStyle", &["title", "ctrTitle"]),
+    ("bodyStyle", &["body", "subTitle", "obj", ""]),
+    ("otherStyle", &["dt", "ftr", "sldNum"]),
+];
+
+/// Parse paragraph alignment from master placeholder shapes' lstStyle > lvl1pPr algn attribute.
+/// Takes the shared, already-parsed master root (see `parse_master_anchors`).
+pub(crate) fn parse_master_alignments(root: roxmltree::Node<'_, '_>) -> HashMap<String, String> {
+    let mut map = HashMap::new();
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            let ph_node = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph");
+            if let Some(ph) = ph_node {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(algn) = child(sp, "txBody")
+                    .and_then(|tb| child(tb, "lstStyle"))
+                    .and_then(|ls| child(ls, "lvl1pPr"))
+                    .and_then(|lp| attr(&lp, "algn"))
+                {
+                    map.entry(ph_type).or_insert(algn);
+                }
+            }
+        }
+    }
+    // Fallback: master <p:txStyles> paragraph alignment (ECMA-376 §19.3.1.52
+    // txStyles → titleStyle §19.3.1.49 / bodyStyle §19.3.1.5 / otherStyle §19.3.1.35).
+    // Per-shape lstStyle (scanned above) wins via or_insert; this fills types
+    // whose master placeholder shape carried no explicit algn (the common case —
+    // PowerPoint stores title/body alignment in txStyles, not the shape lstStyle).
+    if let Some(tx_styles) = child(root, "txStyles") {
+        for &(style, types) in MASTER_TXSTYLE_PH_TYPES {
+            if let Some(algn) = child(tx_styles, style)
+                .and_then(|s| child(s, "lvl1pPr"))
+                .and_then(|lp| attr(&lp, "algn"))
+            {
+                for t in types {
+                    map.entry((*t).to_string()).or_insert_with(|| algn.clone());
+                }
+            }
+        }
+    }
+    map
+}
+
+/// Parse master-level default East Asian line-break (eaLnBrk) per placeholder
+/// type from each placeholder shape's lstStyle > lvl1pPr @eaLnBrk
+/// (ECMA-376 §21.1.2.2.7). Mirrors parse_master_alignments. xsd:boolean.
+pub(crate) fn parse_master_ea_ln_brk(root: roxmltree::Node<'_, '_>) -> HashMap<String, bool> {
+    let mut map = HashMap::new();
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            let ph_node = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph");
+            if let Some(ph) = ph_node {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(v) = child(sp, "txBody")
+                    .and_then(|tb| child(tb, "lstStyle"))
+                    .and_then(|ls| child(ls, "lvl1pPr"))
+                    .and_then(|lp| attr(&lp, "eaLnBrk"))
+                {
+                    map.entry(ph_type).or_insert(v == "1" || v == "true");
+                }
+            }
+        }
+    }
+    map
+}
+
+/// Parse master-level default font sizes from txStyles (titleStyle / bodyStyle / otherStyle)
+/// and from individual placeholder shapes in the master spTree.
+/// Individual shape lstStyle takes priority over txStyles generic defaults.
+pub(crate) fn parse_master_font_sizes(root: roxmltree::Node<'_, '_>) -> HashMap<String, f64> {
+    let mut map = HashMap::new();
+
+    // Scan master spTree placeholder shapes first — per-shape lstStyle is more specific
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            let ph_node = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph");
+            if let Some(ph) = ph_node {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(tx_body) = child(sp, "txBody") {
+                    if let Some(sz) = extract_lvl1_font_size(tx_body) {
+                        map.entry(ph_type).or_insert(sz);
+                    }
+                }
+            }
+        }
+    }
+
+    // p:txStyles > a:titleStyle / a:bodyStyle / a:otherStyle as fallback
+    if let Some(tx_styles) = child(root, "txStyles") {
+        let style_ph_map: &[(&str, &[&str])] = MASTER_TXSTYLE_PH_TYPES;
+        for (style_name, ph_types) in style_ph_map {
+            let sz = child(tx_styles, style_name)
+                .and_then(|sn| child(sn, "lvl1pPr"))
+                .and_then(|lp| child(lp, "defRPr"))
+                .and_then(|rp| attr_f64(&rp, "sz"))
+                .map(|v| v / 100.0);
+            if let Some(fs) = sz {
+                for ph_type in *ph_types {
+                    map.entry(ph_type.to_string()).or_insert(fs);
+                }
+            }
+        }
+    }
+
+    map
+}
+
+/// Per-list-level default font sizes from the master, keyed by ph_type. Mirrors
+/// `parse_master_font_sizes` but captures every list level (lvl1pPr..lvl9pPr) so
+/// nested bullets inherit the correct shrinking sizes (ECMA-376 §21.1.2.4),
+/// not just the level-1 size. Per-shape lstStyle wins over the generic txStyles.
+pub(crate) fn parse_master_level_font_sizes(
+    root: roxmltree::Node<'_, '_>,
+) -> HashMap<String, LevelFontSizes> {
+    let mut map: HashMap<String, LevelFontSizes> = HashMap::new();
+
+    // Per-shape lstStyle first (more specific).
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            if let Some(ph) = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph")
+            {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(tx_body) = child(sp, "txBody") {
+                    let sizes = extract_level_font_sizes(tx_body);
+                    if has_any_level_size(&sizes) {
+                        map.entry(ph_type).or_insert(sizes);
+                    }
+                }
+            }
+        }
+    }
+
+    // txStyles fallback.
+    if let Some(tx_styles) = child(root, "txStyles") {
+        let style_ph_map: &[(&str, &[&str])] = MASTER_TXSTYLE_PH_TYPES;
+        for (style_name, ph_types) in style_ph_map {
+            if let Some(style_node) = child(tx_styles, style_name) {
+                let sizes = read_level_font_sizes(style_node);
+                if has_any_level_size(&sizes) {
+                    for ph_type in *ph_types {
+                        map.entry(ph_type.to_string()).or_insert(sizes);
+                    }
+                }
+            }
+        }
+    }
+
+    map
+}
+
+/// Per-list-level paragraph indents (`marL`/`marR`/`indent`, EMU) from the master,
+/// keyed by ph_type. Mirrors `parse_master_level_font_sizes` exactly (same per-shape
+/// lstStyle then `txStyles` tiers via `MASTER_TXSTYLE_PH_TYPES`): a master body
+/// `<a:lvlNpPr@marL>` is what a slide body paragraph with no own `marL` inherits
+/// (ECMA-376 §21.1.2.4.13). KNOWN SHARED GAP: no presentation `defaultTextStyle`
+/// tier (§19.2.1.8, the lowest authored fallback) — the parser reads it for neither
+/// indents nor font sizes nor bullets, so this stays at parity rather than adding a
+/// tier only here; closing it is a separate cross-cutting change.
+pub(crate) fn parse_master_level_indents(
+    root: roxmltree::Node<'_, '_>,
+) -> HashMap<String, LevelIndents> {
+    let mut map: HashMap<String, LevelIndents> = HashMap::new();
+
+    // Per-shape lstStyle first (more specific).
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            if let Some(ph) = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph")
+            {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(tx_body) = child(sp, "txBody") {
+                    let indents = extract_level_indents(tx_body);
+                    if has_any_level_indent(&indents) {
+                        map.entry(ph_type).or_insert(indents);
+                    }
+                }
+            }
+        }
+    }
+
+    // txStyles fallback.
+    if let Some(tx_styles) = child(root, "txStyles") {
+        let style_ph_map: &[(&str, &[&str])] = MASTER_TXSTYLE_PH_TYPES;
+        for (style_name, ph_types) in style_ph_map {
+            if let Some(style_node) = child(tx_styles, style_name) {
+                let indents = read_level_indents(style_node);
+                if has_any_level_indent(&indents) {
+                    for ph_type in *ph_types {
+                        map.entry(ph_type.to_string()).or_insert(indents);
+                    }
+                }
+            }
+        }
+    }
+
+    map
+}
+
+/// Per-list-level bullets from the master, keyed by ph_type. Mirrors
+/// `parse_master_level_font_sizes`: a master body placeholder's `<a:buChar>` (or
+/// the `bodyStyle` `<a:lvlNpPr>` bullets) is what a slide body paragraph with no
+/// explicit bullet inherits (ECMA-376 §19.7.10 / §21.1.2.4). Per-shape lstStyle
+/// wins over the generic txStyles.
+pub(crate) fn parse_master_level_bullets(
+    root: roxmltree::Node<'_, '_>,
+    theme: &HashMap<String, String>,
+    master_rels: &HashMap<String, String>,
+    master_dir: &str,
+    zip: &mut PptxZip,
+) -> HashMap<String, LevelBullets> {
+    let mut map: HashMap<String, LevelBullets> = HashMap::new();
+
+    // A master-level `<a:buBlip>` embed resolves against the master's rels +
+    // part directory (ECMA-376 §21.1.2.4.2), mirroring the master background.
+    let mut resolve_blip = |rid: &str| -> Option<String> {
+        let target = master_rels.get(rid)?;
+        let path = resolve_path(master_dir, target);
+        // Verify the part exists so a listed-but-missing rId yields None and the
+        // bullet falls through to Bullet::Inherit (matches the variant's doc
+        // comment), mirroring the master background resolver. `index_for_name`
+        // checks the central directory only (no inflate), unlike the former
+        // `read_zip_bytes` which decompressed the entry just to discard it.
+        zip.index_for_name(&path)?;
+        Some(path)
+    };
+
+    // Per-shape lstStyle first (more specific).
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            if let Some(ph) = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph")
+            {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(tx_body) = child(sp, "txBody") {
+                    let bullets = extract_level_bullets(tx_body, theme, &mut resolve_blip);
+                    if has_any_level_bullet(&bullets) {
+                        map.entry(ph_type).or_insert(bullets);
+                    }
+                }
+            }
+        }
+    }
+
+    // txStyles fallback.
+    if let Some(tx_styles) = child(root, "txStyles") {
+        let style_ph_map: &[(&str, &[&str])] = MASTER_TXSTYLE_PH_TYPES;
+        for (style_name, ph_types) in style_ph_map {
+            if let Some(style_node) = child(tx_styles, style_name) {
+                let bullets = read_level_bullets(style_node, theme, &mut resolve_blip);
+                if has_any_level_bullet(&bullets) {
+                    for ph_type in *ph_types {
+                        map.entry(ph_type.to_string())
+                            .or_insert_with(|| bullets.clone());
+                    }
+                }
+            }
+        }
+    }
+
+    map
+}
+
+/// Parse default bold/italic from master txStyles (titleStyle / bodyStyle / otherStyle)
+/// > lvl1pPr > defRPr @b and @i. Keyed by ph_type.
+/// > Only populated when the attribute is explicitly present on the master.
+pub(crate) fn parse_master_txstyle_bold_italic(
+    root: roxmltree::Node<'_, '_>,
+) -> (
+    HashMap<String, bool>,
+    HashMap<String, bool>,
+    HashMap<String, String>,
+) {
+    let mut bold_map: HashMap<String, bool> = HashMap::new();
+    let mut italic_map: HashMap<String, bool> = HashMap::new();
+    // ECMA-376 §21.1.2.3.13 cap="all"/"small" on the master txStyles defRPr —
+    // e.g. a template titleStyle with cap="all" upper-cases every title.
+    let mut caps_map: HashMap<String, String> = HashMap::new();
+    let Some(tx_styles) = child(root, "txStyles") else {
+        return (bold_map, italic_map, caps_map);
+    };
+    let style_ph_map: &[(&str, &[&str])] = MASTER_TXSTYLE_PH_TYPES;
+    for (style_name, ph_types) in style_ph_map {
+        let def_rpr = child(tx_styles, style_name)
+            .and_then(|sn| child(sn, "lvl1pPr"))
+            .and_then(|lp| child(lp, "defRPr"));
+        let b = def_rpr
+            .and_then(|rp| attr(&rp, "b"))
+            .map(|v| v == "1" || v == "true");
+        let i = def_rpr
+            .and_then(|rp| attr(&rp, "i"))
+            .map(|v| v == "1" || v == "true");
+        let c = def_rpr
+            .and_then(|rp| attr(&rp, "cap"))
+            .filter(|v| v == "all" || v == "small");
+        if let Some(bv) = b {
+            for t in *ph_types {
+                bold_map.entry(t.to_string()).or_insert(bv);
+            }
+        }
+        if let Some(iv) = i {
+            for t in *ph_types {
+                italic_map.entry(t.to_string()).or_insert(iv);
+            }
+        }
+        if let Some(cv) = c {
+            for t in *ph_types {
+                caps_map.entry(t.to_string()).or_insert(cv.clone());
+            }
+        }
+    }
+    (bold_map, italic_map, caps_map)
+}
+
+/// Parse default text color from master txStyles (titleStyle/bodyStyle/otherStyle)
+/// > lvl1pPr > defRPr > solidFill, and from per-placeholder shapes in the master spTree's
+/// > txBody > lstStyle > lvl1pPr > defRPr > solidFill. Keyed by ph_type.
+/// > Shape-level lstStyle takes priority over txStyles generic defaults.
+pub(crate) fn parse_master_txstyle_color(
+    root: roxmltree::Node<'_, '_>,
+    theme: &HashMap<String, String>,
+) -> HashMap<String, String> {
+    let mut map: HashMap<String, String> = HashMap::new();
+
+    // Scan master spTree placeholder shapes first — per-shape lstStyle is more specific.
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            let ph_node = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph");
+            if let Some(ph) = ph_node {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(color) = child(sp, "txBody")
+                    .and_then(|tb| child(tb, "lstStyle"))
+                    .and_then(|ls| child(ls, "lvl1pPr"))
+                    .and_then(|lp| child(lp, "defRPr"))
+                    .and_then(|rp| child(rp, "solidFill"))
+                    .and_then(|sf| parse_color_node(sf, theme))
+                {
+                    map.entry(ph_type).or_insert(color);
+                }
+            }
+        }
+    }
+
+    // Fall back to p:txStyles > titleStyle/bodyStyle/otherStyle > lvl1pPr > defRPr > solidFill.
+    if let Some(tx_styles) = child(root, "txStyles") {
+        let style_ph_map: &[(&str, &[&str])] = MASTER_TXSTYLE_PH_TYPES;
+        for (style_name, ph_types) in style_ph_map {
+            if let Some(color) = child(tx_styles, style_name)
+                .and_then(|sn| child(sn, "lvl1pPr"))
+                .and_then(|lp| child(lp, "defRPr"))
+                .and_then(|rp| child(rp, "solidFill"))
+                .and_then(|sf| parse_color_node(sf, theme))
+            {
+                for ph_type in *ph_types {
+                    map.entry(ph_type.to_string()).or_insert(color.clone());
+                }
+            }
+        }
+    }
+
+    map
+}
+
+/// Parse default paragraph spacing from master txStyles.
+/// Returns (space_before_map, space_after_map, line_spacing_map) keyed by ph_type string.
+/// space_before/after values are in hundredths of a point (same as Paragraph.space_before/after).
+/// Note: line_spacing_map is intentionally NOT populated. Inheriting txStyles lnSpc hurts VRT
+/// scores because our font substitutes (sans-serif) have different em-square metrics than the
+/// original Aptos font, so applying the master's 120% line spacing over-expands text layout.
+pub(crate) fn parse_master_txstyle_spacing(
+    root: roxmltree::Node<'_, '_>,
+) -> (
+    HashMap<String, i64>,
+    HashMap<String, i64>,
+    HashMap<String, f64>,
+) {
+    let mut before_map: HashMap<String, i64> = HashMap::new();
+    let mut after_map: HashMap<String, i64> = HashMap::new();
+    let line_map: HashMap<String, f64> = HashMap::new(); // intentionally not populated
+    let tx_styles = match child(root, "txStyles") {
+        Some(n) => n,
+        None => return (before_map, after_map, line_map),
+    };
+    let style_ph_map: &[(&str, &[&str])] = MASTER_TXSTYLE_PH_TYPES;
+    for (style_name, ph_types) in style_ph_map {
+        let lvl1 = child(tx_styles, style_name).and_then(|sn| child(sn, "lvl1pPr"));
+        let spc_before = lvl1
+            .and_then(|lp| child(lp, "spcBef"))
+            .and_then(|s| child(s, "spcPts").and_then(|n| attr_i64(&n, "val")));
+        let spc_after = lvl1
+            .and_then(|lp| child(lp, "spcAft"))
+            .and_then(|s| child(s, "spcPts").and_then(|n| attr_i64(&n, "val")));
+        if let Some(v) = spc_before {
+            for ph_type in *ph_types {
+                before_map.entry(ph_type.to_string()).or_insert(v);
+            }
+        }
+        if let Some(v) = spc_after {
+            for ph_type in *ph_types {
+                after_map.entry(ph_type.to_string()).or_insert(v);
+            }
+        }
+    }
+    (before_map, after_map, line_map)
+}
+
+pub(crate) fn parse_master_transforms(root: roxmltree::Node<'_, '_>) -> HashMap<String, Transform> {
+    let mut map = HashMap::new();
+    if let Some(sp_tree) = child(root, "cSld").and_then(|n| child(n, "spTree")) {
+        for sp in sp_tree
+            .children()
+            .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+        {
+            let ph_node = sp
+                .descendants()
+                .find(|n| n.is_element() && n.tag_name().name() == "ph");
+            if let Some(ph) = ph_node {
+                let ph_type = attr(&ph, "type").unwrap_or_default();
+                if let Some(xfrm) = child(sp, "spPr").and_then(|p| child(p, "xfrm")) {
+                    map.entry(ph_type).or_insert_with(|| parse_xfrm(xfrm));
+                }
+            }
+        }
+    }
+    map
+}
+
+// Seeds layout placeholders from the master's per-type defaults (transforms,
+// alignment, spacing) before overlaying the layout's own placeholder props; the
+// many maps are the master inheritance sources, threaded through as-is.
+//
+// Takes the already-parsed layout root (`<p:sldLayout>`) so `parse_layout` can
+// parse the layout XML once and share the `Document` with the background +
+// showMasterSp extractions (D4).
+#[allow(clippy::too_many_arguments)]
+pub(crate) fn parse_layout_placeholders(
+    root: roxmltree::Node<'_, '_>,
+    master_font_sizes: &HashMap<String, f64>,
+    master_level_font_sizes: &HashMap<String, LevelFontSizes>,
+    master_level_indents: &HashMap<String, LevelIndents>,
+    master_level_bullets: &HashMap<String, LevelBullets>,
+    master_anchors: &HashMap<String, String>,
+    master_transforms: &HashMap<String, Transform>,
+    master_alignments: &HashMap<String, String>,
+    master_ea_ln_brk: &HashMap<String, bool>,
+    master_space_before: &HashMap<String, i64>,
+    master_space_after: &HashMap<String, i64>,
+    master_line_spacing: &HashMap<String, f64>,
+    theme: &HashMap<String, String>,
+    layout_dir: &str,
+    layout_rels: &HashMap<String, String>,
+    zip: &mut PptxZip,
+) -> LayoutPlaceholders {
+    let mut lph = LayoutPlaceholders {
+        master_by_type: master_transforms.clone(),
+        by_type_master_alignment: master_alignments.clone(),
+        by_type_master_ea_ln_brk: master_ea_ln_brk.clone(),
+        by_type_master_space_before: master_space_before.clone(),
+        by_type_master_space_after: master_space_after.clone(),
+        by_type_master_line_spacing: master_line_spacing.clone(),
+        ..Default::default()
+    };
+
+    let sp_tree = root
+        .descendants()
+        .find(|n| n.is_element() && n.tag_name().name() == "spTree");
+    let sp_tree = match sp_tree {
+        Some(n) => n,
+        None => return lph,
+    };
+
+    for sp in sp_tree
+        .children()
+        .filter(|n| n.is_element() && n.tag_name().name() == "sp")
+    {
+        let ph_node = sp
+            .descendants()
+            .find(|n| n.is_element() && n.tag_name().name() == "ph");
+        let sp_pr = match child(sp, "spPr") {
+            Some(n) => n,
+            None => continue,
+        };
+        // xfrm may be absent (placeholder inherits transform from master); parse if present
+        let t_opt: Option<Transform> = child(sp_pr, "xfrm").map(parse_xfrm);
+
+        // Extract layout-level defaults from the placeholder's txBody > lstStyle > lvl1pPr
+        let layout_lvl1_ppr: Option<roxmltree::Node<'_, '_>> = child(sp, "txBody")
+            .and_then(|tb| child(tb, "lstStyle"))
+            .and_then(|ls| child(ls, "lvl1pPr"));
+        let layout_def_rpr: Option<roxmltree::Node<'_, '_>> =
+            layout_lvl1_ppr.and_then(|lp| child(lp, "defRPr"));
+        let layout_font_size = layout_def_rpr
+            .and_then(|rp| attr_f64(&rp, "sz"))
+            .map(|v| v / 100.0);
+        // Per-level sizes from the layout placeholder's own lstStyle (all
+        // lvlNpPr), used to give nested bullets their shrinking sizes.
+        let layout_level_sizes: LevelFontSizes = child(sp, "txBody")
+            .map(extract_level_font_sizes)
+            .unwrap_or([None; 9]);
+        // Per-level indents (marL/marR/indent) from the layout placeholder's own
+        // lstStyle, the inherited list-indent cascade (ECMA-376 §21.1.2.4.13).
+        let layout_level_indents: LevelIndents = child(sp, "txBody")
+            .map(extract_level_indents)
+            .unwrap_or_default();
+        // Per-level bullets from the layout placeholder's own lstStyle. A
+        // level's `<a:buBlip>` embed (§21.1.2.4.2) resolves against the layout's
+        // rels + part directory, mirroring the layout-spPr blipFill above.
+        let mut resolve_layout_blip = |rid: &str| -> Option<String> {
+            let target = layout_rels.get(rid)?;
+            let path = resolve_path(layout_dir, target);
+            // Verify the part exists so a listed-but-missing rId yields None and
+            // the bullet falls through to Bullet::Inherit (matches the variant's
+            // doc comment), mirroring the master/layout background resolvers.
+            // `index_for_name` reads the central directory only (no inflate),
+            // unlike the former `read_zip_bytes` which decompressed and discarded.
+            zip.index_for_name(&path)?;
+            Some(path)
+        };
+        let layout_level_bullets: LevelBullets = child(sp, "txBody")
+            .map(|tb| extract_level_bullets(tb, theme, &mut resolve_layout_blip))
+            .unwrap_or_else(empty_level_bullets);
+        let layout_bold = layout_def_rpr
+            .and_then(|rp| attr(&rp, "b"))
+            .map(|v| v == "1" || v == "true");
+        let layout_italic = layout_def_rpr
+            .and_then(|rp| attr(&rp, "i"))
+            .map(|v| v == "1" || v == "true");
+        let layout_caps = layout_def_rpr
+            .and_then(|rp| attr(&rp, "cap"))
+            .filter(|v| v == "all" || v == "small");
+        let layout_color: Option<String> = layout_def_rpr
+            .and_then(|rp| child(rp, "solidFill"))
+            .and_then(|sf| parse_color_node(sf, theme));
+        let layout_alignment: Option<String> = layout_lvl1_ppr
+            .and_then(|lp| attr(&lp, "algn"))
+            .map(|a| a.to_string());
+        // ECMA-376 §21.1.2.2.7 eaLnBrk from the layout placeholder's lvl1pPr.
+        let layout_ea_ln_brk: Option<bool> = layout_lvl1_ppr
+            .and_then(|lp| attr(&lp, "eaLnBrk"))
+            .map(|v| v == "1" || v == "true");
+        let layout_space_before: Option<i64> = layout_lvl1_ppr
+            .and_then(|lp| child(lp, "spcBef"))
+            .and_then(|s| child(s, "spcPts"))
+            .and_then(|s| attr_i64(&s, "val"));
+        let layout_space_after: Option<i64> = layout_lvl1_ppr
+            .and_then(|lp| child(lp, "spcAft"))
+            .and_then(|s| child(s, "spcPts"))
+            .and_then(|s| attr_i64(&s, "val"));
+        // lnSpc > spcPct val (e.g. 90000 = 90%)
+        let layout_line_spacing: Option<f64> = layout_lvl1_ppr
+            .and_then(|lp| child(lp, "lnSpc"))
+            .and_then(|ls| child(ls, "spcPct"))
+            .and_then(|s| attr_f64(&s, "val"));
+
+        // Layout bodyPr anchor; fall back to master anchor map
+        let layout_anchor: Option<String> = child(sp, "txBody")
+            .and_then(|tb| child(tb, "bodyPr"))
+            .and_then(|bp| attr(&bp, "anchor"))
+            .map(|a| a.to_string());
+
+        // Layout spPr > ln stroke (real visible border, not edit-mode indicator when solidFill is present)
+        let layout_stroke: Option<Stroke> = child(sp_pr, "ln").and_then(|n| parse_stroke(n, theme));
+
+        // Layout spPr fill (solidFill / noFill / gradFill / pattFill). The
+        // slide-level placeholder shape inherits this when its own `<p:spPr>` is
+        // empty — that's how a "tinted body placeholder" carries through to the
+        // slide. We deliberately exclude grpFill here (group inheritance is
+        // resolved at slide parse time, not from the layout).
+        let layout_fill: Option<Fill> = parse_fill(sp_pr, theme);
+
+        // Layout spPr > blipFill → image that bleeds through when the slide's
+        // matching placeholder has no own blipFill (picture placeholder inheritance).
+        let layout_blip_fill: Option<InheritedBlipFill> = child(sp_pr, "blipFill").and_then(|bf| {
+            let rid = child(bf, "blip").and_then(|b| attr_r(&b, "embed"))?;
+            let rel_target = layout_rels.get(&rid)?;
+            let image_path = resolve_path(layout_dir, rel_target);
+            // Verify the part exists so a dangling rId yields None (no inherited
+            // fill), preserving the prior data-URL behaviour. `index_for_name`
+            // reads the central directory only (no inflate), unlike the former
+            // `read_zip_bytes` which decompressed the entry just to discard it.
+            zip.index_for_name(&image_path)?;
+            let mime_type = mime_from_ext(&image_path).to_owned();
+            Some(InheritedBlipFill {
+                image_path,
+                mime_type,
+                src_rect: parse_src_rect(bf),
+                alpha: parse_blip_alpha(bf),
+            })
+        });
+
+        if let Some(ph) = ph_node {
+            let ph_type = attr(&ph, "type").unwrap_or_default();
+            let ph_idx: Option<u32> = attr(&ph, "idx").and_then(|v| v.parse().ok());
+
+            if let Some(idx) = ph_idx {
+                if let Some(ref t) = t_opt {
+                    lph.by_idx.entry(idx).or_insert_with(|| t.clone());
+                }
+                // Prefer layout font size; fall back to master
+                let fs = layout_font_size.or_else(|| master_font_sizes.get(&ph_type).copied());
+                if let Some(fs) = fs {
+                    lph.by_idx_font_size.entry(idx).or_insert(fs);
+                }
+                // Per-level: layout lstStyle wins per level, else master.
+                let level_sizes = merge_level_sizes(
+                    &layout_level_sizes,
+                    master_level_font_sizes.get(&ph_type).unwrap_or(&[None; 9]),
+                );
+                if has_any_level_size(&level_sizes) {
+                    lph.by_idx_level_sizes.entry(idx).or_insert(level_sizes);
+                }
+                // Per-level indents: layout lstStyle wins per axis/level, else master.
+                let level_indents = merge_level_indents(
+                    &layout_level_indents,
+                    master_level_indents
+                        .get(&ph_type)
+                        .unwrap_or(&Default::default()),
+                );
+                if has_any_level_indent(&level_indents) {
+                    lph.by_idx_level_indents.entry(idx).or_insert(level_indents);
+                }
+                // Per-level bullets: layout lstStyle wins per level, else master.
+                let empty_bul = empty_level_bullets();
+                let level_bullets = merge_level_bullets(
+                    &layout_level_bullets,
+                    master_level_bullets.get(&ph_type).unwrap_or(&empty_bul),
+                );
+                if has_any_level_bullet(&level_bullets) {
+                    lph.by_idx_level_bullets.entry(idx).or_insert(level_bullets);
+                }
+                if let Some(ref s) = layout_stroke {
+                    lph.by_idx_stroke.entry(idx).or_insert(s.clone());
+                }
+                if let Some(ls) = layout_line_spacing {
+                    lph.by_idx_line_spacing.entry(idx).or_insert(ls);
+                }
+                if let Some(ref bf) = layout_blip_fill {
+                    lph.by_idx_blip_fill.entry(idx).or_insert(bf.clone());
+                }
+                if let Some(ref c) = layout_color {
+                    lph.by_idx_color.entry(idx).or_insert(c.clone());
+                }
+                if let Some(ref f) = layout_fill {
+                    lph.by_idx_fill.entry(idx).or_insert(f.clone());
+                }
+                // Alignment for this idx: layout's own algn, else master per-type
+                // (incl. master txStyles, now folded into master_alignments).
+                let idx_algn = layout_alignment
+                    .clone()
+                    .or_else(|| master_alignments.get(&ph_type).cloned());
+                if let Some(a) = idx_algn {
+                    lph.by_idx_alignment.entry(idx).or_insert(a);
+                }
+            }
+            let effective_fs =
+                layout_font_size.or_else(|| master_font_sizes.get(&ph_type).copied());
+            if let Some(fs) = effective_fs {
+                lph.by_type_font_size.entry(ph_type.clone()).or_insert(fs);
+            }
+            let type_level_sizes = merge_level_sizes(
+                &layout_level_sizes,
+                master_level_font_sizes.get(&ph_type).unwrap_or(&[None; 9]),
+            );
+            if has_any_level_size(&type_level_sizes) {
+                lph.by_type_level_sizes
+                    .entry(ph_type.clone())
+                    .or_insert(type_level_sizes);
+            }
+            let type_level_indents = merge_level_indents(
+                &layout_level_indents,
+                master_level_indents
+                    .get(&ph_type)
+                    .unwrap_or(&Default::default()),
+            );
+            if has_any_level_indent(&type_level_indents) {
+                lph.by_type_level_indents
+                    .entry(ph_type.clone())
+                    .or_insert(type_level_indents);
+            }
+            let empty_bul_t = empty_level_bullets();
+            let type_level_bullets = merge_level_bullets(
+                &layout_level_bullets,
+                master_level_bullets.get(&ph_type).unwrap_or(&empty_bul_t),
+            );
+            if has_any_level_bullet(&type_level_bullets) {
+                lph.by_type_level_bullets
+                    .entry(ph_type.clone())
+                    .or_insert(type_level_bullets);
+            }
+            if let Some(b) = layout_bold {
+                lph.by_type_bold.entry(ph_type.clone()).or_insert(b);
+            }
+            if let Some(i) = layout_italic {
+                lph.by_type_italic.entry(ph_type.clone()).or_insert(i);
+            }
+            if let Some(c) = layout_caps.clone() {
+                lph.by_type_caps.entry(ph_type.clone()).or_insert(c);
+            }
+            if let Some(a) = layout_alignment {
+                lph.by_type_alignment.entry(ph_type.clone()).or_insert(a);
+            }
+            if let Some(e) = layout_ea_ln_brk {
+                lph.by_type_ea_ln_brk.entry(ph_type.clone()).or_insert(e);
+            }
+            if let Some(v) = layout_space_before {
+                lph.by_type_space_before.entry(ph_type.clone()).or_insert(v);
+            }
+            if let Some(v) = layout_space_after {
+                lph.by_type_space_after.entry(ph_type.clone()).or_insert(v);
+            }
+            if let Some(ls) = layout_line_spacing {
+                lph.by_type_line_spacing
+                    .entry(ph_type.clone())
+                    .or_insert(ls);
+            }
+            // Anchor: layout bodyPr > fall back to master anchor map
+            let effective_anchor = layout_anchor
+                .clone()
+                .or_else(|| master_anchors.get(&ph_type).cloned());
+            if let Some(a) = effective_anchor {
+                lph.by_type_anchor.entry(ph_type.clone()).or_insert(a);
+            }
+            if let Some(s) = layout_stroke {
+                lph.by_type_stroke.entry(ph_type.clone()).or_insert(s);
+            }
+            if let Some(bf) = layout_blip_fill {
+                lph.by_type_blip_fill.entry(ph_type.clone()).or_insert(bf);
+            }
+            if let Some(c) = layout_color {
+                lph.by_type_color.entry(ph_type.clone()).or_insert(c);
+            }
+            if let Some(f) = layout_fill {
+                lph.by_type_fill.entry(ph_type.clone()).or_insert(f);
+            }
+            if let Some(t) = t_opt {
+                lph.by_type.entry(ph_type).or_insert(t);
+            }
+        }
+    }
+    lph
+}
+
+/// The layout XML parsed ONCE into the owned data a slide needs from its layout
+/// (D4). Groups the three former per-slide layout re-parses in `parse_slide`:
+/// placeholder inheritance (§19.3.1.39), the layout-level `<p:bg>` background,
+/// and the layout's `showMasterSp` flag (§19.3.1.39). Holds no `roxmltree` node
+/// (owned only), so it can be cached across slides sharing a layout.
+///
+/// The color-bearing fields (`placeholders` colors/fills/strokes/bullets +
+/// `background`) are resolved against the `theme` passed to `parse_layout`. For
+/// the common no-`clrMapOvr` slide that theme is the master's baked theme, so
+/// the cached instance is reused; a slide with a `<p:clrMapOvr>` builds a fresh
+/// `ParsedLayout` against its override theme (see the `parse_presentation` loop)
+/// so its layout colors flip too. The layout's DECORATIVE spTree shapes are NOT
+/// held here — they are walked per-slide because they resolve against the slide's
+/// own `smartart_drawings` (§19.3.1.39 layout decorations) and are theme+zip
+/// bound; caching them keyed by layout would be unsound.
+pub(crate) struct ParsedLayout {
+    pub(crate) placeholders: LayoutPlaceholders,
+    /// Layout-level `<p:cSld><p:bg>` fill (ECMA-376 §19.3.1.1 / §20.1.8.14),
+    /// resolved against `theme`. Applied by the slide only when its own bg chain
+    /// (slide-level) resolves to nothing.
+    pub(crate) background: Option<Fill>,
+    /// The LAYOUT's own `showMasterSp` (§19.3.1.39). The slide ANDs this with its
+    /// own slide-level flag before compositing master decorations.
+    pub(crate) show_master_sp: bool,
+}
+
+impl Default for ParsedLayout {
+    fn default() -> Self {
+        // Matches the prior "no/unparseable layout" behaviour: no placeholders,
+        // no layout background, and showMasterSp defaulting to true.
+        ParsedLayout {
+            placeholders: LayoutPlaceholders::default(),
+            background: None,
+            show_master_sp: true,
+        }
+    }
+}
+
+/// ECMA-376 §19.3.1.38/§19.3.1.39 showMasterSp: absent / "1" / "true" ⇒ true;
+/// "0" / "false" ⇒ false. Read from a slide or layout root element.
+pub(crate) fn read_show_master_sp(node: roxmltree::Node<'_, '_>) -> bool {
+    match attr(&node, "showMasterSp").as_deref() {
+        Some("0") | Some("false") => false,
+        _ => true, // default true (absent / "1" / "true")
+    }
+}
+
+/// Parse a slide layout's XML EXACTLY ONCE and extract everything a slide
+/// inherits from it (D4). Replaces the four former per-slide layout
+/// `Document::parse` calls in `parse_slide` (placeholders, background,
+/// showMasterSp, decorations) — the decorations still walk per-slide, but from
+/// the SAME `Document` when the caller reuses it, and the other three are cached.
+/// `theme` is the slide's effective theme (master-baked, or override-adjusted);
+/// the master maps are the inheritance fallbacks, threaded through unchanged.
+#[allow(clippy::too_many_arguments)]
+pub(crate) fn parse_layout(
+    layout_xml: &str,
+    master_font_sizes: &HashMap<String, f64>,
+    master_level_font_sizes: &HashMap<String, LevelFontSizes>,
+    master_level_indents: &HashMap<String, LevelIndents>,
+    master_level_bullets: &HashMap<String, LevelBullets>,
+    master_anchors: &HashMap<String, String>,
+    master_transforms: &HashMap<String, Transform>,
+    master_alignments: &HashMap<String, String>,
+    master_ea_ln_brk: &HashMap<String, bool>,
+    master_space_before: &HashMap<String, i64>,
+    master_space_after: &HashMap<String, i64>,
+    master_line_spacing: &HashMap<String, f64>,
+    theme: &HashMap<String, String>,
+    layout_dir: &str,
+    layout_rels: &HashMap<String, String>,
+    zip: &mut PptxZip,
+) -> ParsedLayout {
+    note_layout_master_parse();
+    let doc = match roxmltree::Document::parse(layout_xml) {
+        Ok(d) => d,
+        // Unparseable layout → same as no layout: default placeholders/bg and
+        // showMasterSp = true (the slide's own flag still applies downstream).
+        Err(_) => return ParsedLayout::default(),
+    };
+    let root = doc.root_element();
+
+    let placeholders = parse_layout_placeholders(
+        root,
+        master_font_sizes,
+        master_level_font_sizes,
+        master_level_indents,
+        master_level_bullets,
+        master_anchors,
+        master_transforms,
+        master_alignments,
+        master_ea_ln_brk,
+        master_space_before,
+        master_space_after,
+        master_line_spacing,
+        theme,
+        layout_dir,
+        layout_rels,
+        zip,
+    );
+
+    // Layout-level bg (rels = layout rels, part dir = layout_dir). Verbatim from
+    // the former inline layout-bg block in `parse_slide`; the slide decides
+    // whether to use it (only when its own bg chain is empty).
+    let background: Option<Fill> = child(root, "cSld").and_then(|n| {
+        let mut resolve = |rid: &str| -> Option<String> {
+            let target = layout_rels.get(rid)?;
+            let path = resolve_path(layout_dir, target);
+            // Existence check only — central-directory lookup, no inflate.
+            zip.index_for_name(&path)?;
+            Some(path)
+        };
+        parse_background(n, theme, &mut resolve)
+    });
+
+    let show_master_sp = read_show_master_sp(root);
+
+    ParsedLayout {
+        placeholders,
+        background,
+        show_master_sp,
+    }
+}
+
+/// All slide-master-derived data plus the master's effective theme, bundled so
+/// it can be computed once per master and reused across every slide that shares
+/// that master (ECMA-376 §19.3.1.42 — a deck may have multiple masters, each
+/// with its own theme/clrMap). Resolving theme/master per slide via the
+/// slide→slideLayout→slideMaster→theme rels chain is required so that scheme
+/// colors (e.g. `<a:schemeClr val="accent1">`) pick the right palette.
+pub(crate) struct ParsedMaster {
+    /// The master's effective theme palette, with the master's `<p:clrMap>`
+    /// pre-baked (logical names → slot hex). Includes font/line/objectDefault
+    /// keys exactly as `parse_theme_colors` produced them.
+    pub(crate) theme: HashMap<String, String>,
+    pub(crate) master_xml: Option<String>,
+    pub(crate) master_rels: HashMap<String, String>,
+    pub(crate) master_dir: String,
+    pub(crate) master_smartart_drawings: HashMap<String, String>,
+    pub(crate) master_bg: Option<Fill>,
+    /// The master's own decorative (non-placeholder) spTree shapes, resolved ONCE
+    /// against the master's baked `theme` (§19.3.1.38 showMasterSp). Each slide
+    /// composites these beneath its content; pre-extracting here (per cached
+    /// master) removes the per-slide master-XML re-parse + spTree re-walk (D4).
+    /// A slide with a `<p:clrMapOvr>` re-resolves them against its override theme
+    /// (see `parse_slide`), so these frozen-against-master-theme elements are used
+    /// only by the common no-override slides.
+    pub(crate) master_decorative: Vec<SlideElement>,
+    pub(crate) master_font_sizes: HashMap<String, f64>,
+    pub(crate) master_level_font_sizes: HashMap<String, LevelFontSizes>,
+    pub(crate) master_level_indents: HashMap<String, LevelIndents>,
+    pub(crate) master_level_bullets: HashMap<String, LevelBullets>,
+    pub(crate) master_anchors: HashMap<String, String>,
+    pub(crate) master_transforms: HashMap<String, Transform>,
+    pub(crate) master_alignments: HashMap<String, String>,
+    pub(crate) master_ea_ln_brk: HashMap<String, bool>,
+    pub(crate) master_space_before: HashMap<String, i64>,
+    pub(crate) master_space_after: HashMap<String, i64>,
+    pub(crate) master_line_spacing: HashMap<String, f64>,
+    pub(crate) master_bold: HashMap<String, bool>,
+    pub(crate) master_italic: HashMap<String, bool>,
+    pub(crate) master_caps: HashMap<String, String>,
+    pub(crate) master_color: HashMap<String, String>,
+}
+
+/// The subset of `ParsedMaster` fields that are THEME-DEPENDENT, recomputed for a
+/// slide whose `<p:clrMapOvr><a:overrideClrMapping>` (ECMA-376 §19.3.1.7) replaces
+/// the master's color mapping for the WHOLE slide (§20.1.6.8). `build_master_bundle`
+/// freezes these against the MASTER's own clrMap-baked theme; for an override slide
+/// we re-resolve them against the slide's effective mapping so that master-INHERITED
+/// scheme colors (a `<p:bg>` schemeClr, master txStyles placeholder colors, master
+/// bullet colors) flip together with the slide's own shapes. Owns all its data and
+/// holds no `zip` borrow, so it can be built before `parse_slide(zip)` is called.
+pub(crate) struct EffectiveMaster {
+    /// `bundle.theme` clone with the override mapping applied (logical → slot hex).
+    pub(crate) theme: HashMap<String, String>,
+    /// Master `<p:bg>` re-resolved against `theme` (replaces `ParsedMaster.master_bg`).
+    pub(crate) master_bg: Option<Fill>,
+    /// Master txStyles placeholder colors re-resolved against `theme`.
+    pub(crate) master_color: HashMap<String, String>,
+    /// Master per-level bullet colors re-resolved against `theme`.
+    pub(crate) master_level_bullets: HashMap<String, LevelBullets>,
+}
+
+/// Build a `ParsedMaster` for the master at `master_path` (a ZIP path such as
+/// `ppt/slideMasters/slideMaster2.xml`). Reads the master XML + its rels,
+/// resolves the master's own `/theme` relationship, parses the theme colors,
+/// bakes the master's `<p:clrMap>`, then computes every master-derived map.
+///
+/// `fallback_theme` is the presentation-level theme used only when the master
+/// has no `/theme` relationship of its own (keeps simple single-theme decks and
+/// malformed packages working).
+///
+/// TODO: themeOverride (slide/layout `/themeOverride`, ECMA-376 §14.2.7) is not
+/// yet honored — overrides on the layout or slide would replace parts of the
+/// master theme. Out of scope for per-slide master resolution.
+pub(crate) fn build_master_bundle(
+    master_path: &str,
+    fallback_theme: &HashMap<String, String>,
+    zip: &mut PptxZip,
+) -> ParsedMaster {
+    let master_xml_opt: Option<String> = read_zip_str(zip, master_path).ok();
+
+    let master_dir: String = master_path
+        .rsplit_once('/')
+        .map(|(dir, _)| dir.to_owned())
+        .unwrap_or_else(|| "ppt/slideMasters".to_owned());
+
+    // Master rels: `<master_dir>/_rels/<file>.rels`.
+    let master_file = master_path
+        .split('/')
+        .next_back()
+        .unwrap_or("slideMaster1.xml");
+    let master_rels_xml: String = {
+        let rels_p = format!("{master_dir}/_rels/{master_file}.rels");
+        read_zip_str(zip, &rels_p).unwrap_or_default()
+    };
+    let master_rels: HashMap<String, String> = parse_rels(&master_rels_xml);
+
+    // The master's own theme (slide→…→slideMaster→theme). Fall back to the
+    // presentation theme when the master declares no /theme relationship.
+    let theme_path: Option<String> =
+        find_rel_target_by_type(&master_rels_xml, "/theme").map(|t| resolve_path(&master_dir, &t));
+    let mut theme: HashMap<String, String> = match theme_path
+        .as_deref()
+        .and_then(|p| read_zip_str(zip, p).ok())
+    {
+        Some(theme_xml) => parse_theme_colors(&theme_xml),
+        None => fallback_theme.clone(),
+    };
+    // Bake the master's <p:clrMap> logical-name → slot mapping into the theme.
+    bake_clr_map(&mut theme, master_xml_opt.as_deref());
+
+    let master_smartart_drawings: HashMap<String, String> =
+        build_smartart_drawings(&master_rels_xml, zip);
+
+    // Parse the master XML EXACTLY ONCE and share the resulting `Document` across
+    // every master-derived extractor below (D4: previously each `parse_master_*`
+    // re-ran `Document::parse` on the same string, so a single master cost 12
+    // parses — 11 extractors + the background). The `Document` borrows
+    // `master_xml_opt`, so it lives only for the extraction scope; all owned maps
+    // are computed before it is dropped. When the master has no XML (missing part)
+    // every map defaults to empty, exactly as the prior `Option::map` chain did.
+    let master_doc: Option<roxmltree::Document<'_>> = master_xml_opt.as_deref().and_then(|xml| {
+        note_layout_master_parse();
+        roxmltree::Document::parse(xml).ok()
+    });
+    let master_root: Option<roxmltree::Node<'_, '_>> =
+        master_doc.as_ref().map(|d| d.root_element());
+
+    let master_bg: Option<Fill> = master_root.and_then(|root| {
+        let c_sld = child(root, "cSld")?;
+        let mut resolve = |rid: &str| -> Option<String> {
+            let target = master_rels.get(rid)?;
+            let path = resolve_path(&master_dir, target);
+            // Existence check only — central-directory lookup, no inflate
+            // (former `read_zip_bytes` decompressed the entry just to discard it).
+            zip.index_for_name(&path)?;
+            Some(path)
+        };
+        parse_background(c_sld, &theme, &mut resolve)
+    });
+
+    let master_font_sizes = master_root.map(parse_master_font_sizes).unwrap_or_default();
+    let master_level_font_sizes = master_root
+        .map(parse_master_level_font_sizes)
+        .unwrap_or_default();
+    let master_level_indents = master_root
+        .map(parse_master_level_indents)
+        .unwrap_or_default();
+    let master_level_bullets = master_root
+        .map(|root| parse_master_level_bullets(root, &theme, &master_rels, &master_dir, zip))
+        .unwrap_or_default();
+    let master_anchors = master_root.map(parse_master_anchors).unwrap_or_default();
+    let master_transforms = master_root.map(parse_master_transforms).unwrap_or_default();
+    let master_alignments = master_root.map(parse_master_alignments).unwrap_or_default();
+    let master_ea_ln_brk = master_root.map(parse_master_ea_ln_brk).unwrap_or_default();
+    let (master_space_before, master_space_after, master_line_spacing) = master_root
+        .map(parse_master_txstyle_spacing)
+        .unwrap_or_default();
+    let (master_bold, master_italic, master_caps) = master_root
+        .map(parse_master_txstyle_bold_italic)
+        .unwrap_or_default();
+    let master_color = master_root
+        .map(|root| parse_master_txstyle_color(root, &theme))
+        .unwrap_or_default();
+
+    // Pre-extract the master's decorative (non-placeholder) spTree shapes ONCE,
+    // resolved against the master's baked `theme`. Each slide clones these instead
+    // of re-parsing the master XML and re-walking its spTree (D4; former
+    // per-slide `parse_slide` inline walk). Uses the same shared `master_root` and
+    // the master's own rels + smartart drawings, exactly as the old inline walk did.
+    let mut master_decorative: Vec<SlideElement> = Vec::new();
+    if let Some(root) = master_root {
+        extract_decorative_shapes(
+            root,
+            &master_dir,
+            &master_rels,
+            &master_smartart_drawings,
+            &theme,
+            zip,
+            &mut master_decorative,
+        );
+    }
+
+    ParsedMaster {
+        theme,
+        master_xml: master_xml_opt,
+        master_rels,
+        master_dir,
+        master_smartart_drawings,
+        master_bg,
+        master_decorative,
+        master_font_sizes,
+        master_level_font_sizes,
+        master_level_indents,
+        master_level_bullets,
+        master_anchors,
+        master_transforms,
+        master_alignments,
+        master_ea_ln_brk,
+        master_space_before,
+        master_space_after,
+        master_line_spacing,
+        master_bold,
+        master_italic,
+        master_caps,
+        master_color,
+    }
+}
